fix getter and mutator api
